OS/2: Synchronize EMX code with recent Windows NT (Cygwin) changes.
This commit contains a lot of changes in many EMX-related sources for several years.
Unfortunately, there are not many notes on the changes from the author so it's really
too much of work to split them in smaller pieces.

Also note that this patchset was brushed up by the committer (several questionable changes
and leftovers removed, several minor issues are fixed).